Add shared TTL cache (lib/cache.ts); adopt in rawg-api + ai-game-service
Extract the duplicated { timestamp, data } localStorage cache into one
SSR-safe helper (getCache/setCache/clearCache). Adopt it in rawg-api and
ai-game-service, removing their bespoke copies. Stored shape preserved so
existing cached entries stay valid.
